Further code clean up, rename NYPLWelcomeAccountList to AccountList
